Update cors.md
Hi Team, I found a bit of unnatural sentences in Japanese language, and I would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you would check this proposal. Please let me know if you have any concerns.

This request is in regard to:

- https://docs.microsoft.com/ja-jp/aspnet/core/security/cors
- https://docs.microsoft.com/en-us/aspnet/core/security/cors

Best regards,